Working on cleaning up code on the main package of the server
