+++ server/cmd/api/response_messages.go
@@ -3,4 +3,5 @@ package main
 const (
 	internalServerErrorMessage = "The server encountered an unexpected error"
 	notAuthorizedMessage       = "not authorized"
+	invalidRequestBodyMessage  = "invalid request body"
 )
