+++ server/cmd/api/tmdb_handler.go
@@ -68,7 +68,7 @@ func (app *Application) handleTmdbMovieSearch(w http.ResponseWriter, r *http.Req
 
 	err := helpers.ReadJSON(w, r, &payload, 0)
 	if err != nil {
-		helpers.ErrorJSON(w, errors.New("invalid request body"), http.StatusBadRequest)
+		helpers.ErrorJSON(w, errors.New(invalidRequestBodyMessage), http.StatusBadRequest)
 		return
 	}
