+++ server/cmd/api/track_handler.go
@@ -14,9 +14,8 @@ import (
 )
 
 func (app *Application) ToggleLikeTrack(w http.ResponseWriter, r *http.Request) {
-	userID := app.userIDFromRequest(r)
-	if userID == 0 {
-		helpers.ErrorJSON(w, errors.New(notAuthorizedMessage), http.StatusUnauthorized)
+	userID, ok := app.currentUserID(w, r)
+	if !ok {
 		return
 	}
 
@@ -87,9 +86,8 @@ func (app *Application) ToggleLikeTrack(w http.ResponseWriter, r *http.Request) 
 // GetLikedTrackIDsForUser returns the IDs of every track the authenticated user has liked.
 // Used by the frontend to initialize liked-heart state on page load without a paginated fetch.
 func (app *Application) GetLikedTrackIDsForUser(w http.ResponseWriter, r *http.Request) {
-	userID := app.userIDFromRequest(r)
-	if userID == 0 {
-		helpers.ErrorJSON(w, errors.New(notAuthorizedMessage), http.StatusUnauthorized)
+	userID, ok := app.currentUserID(w, r)
+	if !ok {
 		return
 	}
 
@@ -113,9 +111,8 @@ func (app *Application) GetLikedTrackIDsForUser(w http.ResponseWriter, r *http.R
 // GetLikedTracks returns the current user's liked tracks as a paginated list,
 // ordered by most recently liked first. Query params: page (default 1), per_page (default 50, max 100).
 func (app *Application) GetLikedTracks(w http.ResponseWriter, r *http.Request) {
-	userID := app.userIDFromRequest(r)
-	if userID == 0 {
-		helpers.ErrorJSON(w, errors.New(notAuthorizedMessage), http.StatusUnauthorized)
+	userID, ok := app.currentUserID(w, r)
+	if !ok {
 		return
 	}
